Fixes minor compiler warnings related to size_t conversions

+++ src/sqlcipher.c
@@ -1830,7 +1830,7 @@ void sqlcipher_log(unsigned int level, unsigned int source, const char *message,
   va_list params;
   va_start(params, message);
   char formatted[MAX_LOG_LEN];
-  int len = 0;
+  size_t len = 0;
 
 #ifdef CODEC_DEBUG
 #if defined(SQLCIPHER_OMIT_LOG_DEVICE) || (!defined(__ANDROID__) && !defined(__APPLE__))
@@ -1859,7 +1859,7 @@ void sqlcipher_log(unsigned int level, unsigned int source, const char *message,
 
   sqlite3_snprintf(MAX_LOG_LEN, formatted, "%s %s ", sqlcipher_get_log_level_str(level), sqlcipher_get_log_source_str(source));
   len = strlen(formatted);
-  sqlite3_vsnprintf(MAX_LOG_LEN - len, formatted + len, message, params);
+  sqlite3_vsnprintf(MAX_LOG_LEN - (int) len, formatted + (int) len, message, params);
 
 #if !defined(SQLCIPHER_OMIT_LOG_DEVICE)
   if(sqlcipher_log_device) {
